25.04.03.(목) 10:03

+++ src/main/java/com/solomarket/dto/ProductDto.java
@@ -17,8 +17,8 @@ import java.time.LocalDateTime;
  */
 @Data
 public class ProductDto {
-    private int productNo;
-    private int sellerId;
+    private Long productNo;
+    private Long sellerId;
     private String title;
     private int price;
     private String content;
